membuat login midleware dan pelunasan hutang

+++ database/migrations/2018_05_13_110331_create_barangs_table.php
@@ -20,7 +20,7 @@ class CreateBarangsTable extends Migration
             $table->string('nama_barang',100);
             $table->integer('harga_beli');
             $table->integer('harga_jual');
-            $table->integer('stok');
+            $table->float('stok',8,2);
             $table->enum('status_barang',['Publish','Draft'])->default('Publish');
             $table->timestamps();
